Change exported file names from download.csv to descriptive names
Resolves #390.

Related:
http://stackoverflow.com/questions/32653448/change-file-name-when-exporting-to-csv-in-angular-ui-grid

Instead of using the using the exporter function from the scope
we want to be able to change the behavior of the csvExport
when exporting from the grid menu. This requires overriding the
service's csvExport function:

https://github.com/angular-ui/ui-grid/blob/68681d71823fc0c1b9f45e9fa7832de71233c7ab/src/features/exporter/js/exporter.js#L622

If gridOptions has a csvFileNameFn, then that function is called and
used as the filename, otherwise default to original behavior.