Make grid columns resizable.
Add header tooltips to each column definition since
it is possible to resize the columns to make the
headers unreadable.

This allows the user to choose resize the width of
each column in the grids. It may get irritating to
constantly have to resize the columns when navigating
away some we probably want to implement the ui-grid's
saveState feature at some point.